+++ neoforge/src/main/java/com/st0x0ef/stellaris/neoforge/client/StellarisNeoforgeClient.java
@@ -132,10 +132,8 @@ public class StellarisNeoforgeClient {
 
     @SubscribeEvent
     public static void registerKeyBindings(RegisterKeyMappingsEvent event) {
-        event.register(KeyMappingsRegistry.ROCKET_START);
         event.register(KeyMappingsRegistry.CHANGE_JETSUIT_MODE);
         event.register(KeyMappingsRegistry.FREEZE_PLANET_MENU);
-        event.register(KeyMappingsRegistry.SLOW_LANDER);
 
         NeoForge.EVENT_BUS.addListener(StellarisNeoforgeClient::clientTick);
     }