+++ changelog.md
@@ -48,8 +48,10 @@
 - GUI textures for all machines
 - Optimization of the oxygen system
 - Oxygen distributor now propagate in all directions
+- Improve achievement screen
 
 ## Fix
 - Crash when trying to launch to any planet when you haven't launched from a planet
+- Remove configurable start rocket key and slow down lander key to fix an issue with jumping